Renamed imagescraper & imagescraper saves images to a folder
